Moved some code around, some minor renamings
